+++ src/components/format-bytes/format-bytes.ts
@@ -1,9 +1,9 @@
 import { LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
-import { formatBytes } from '~/internal/number';
 import { LocalizeController } from '~/utilities/localize';
 
 /**
+
  * @since 2.0
  * @status stable
  */
@@ -15,15 +15,30 @@ export default class SlFormatBytes extends LitElement {
   @property({ type: Number }) value = 0;
 
   /** The unit to display. */
-  @property() unit: 'bytes' | 'bits' = 'bytes';
+  @property() unit: 'byte' | 'bit' = 'byte';
+
+  /** Determines how to display the result, e.g. "100 bytes", "100 b", or "100b". */
+  @property() display: 'long' | 'short' | 'narrow' = 'short';
 
   /** The locale to use when formatting the number. */
   @property() lang: string;
 
   render() {
-    return formatBytes(this.value, {
-      unit: this.unit,
-      formatter: num => this.localize.number(num)
+    if (isNaN(this.value)) {
+      return '';
+    }
+
+    const bitPrefixes = ['', 'kilo', 'mega', 'giga', 'tera']; // petabit isn't a supported unit
+    const bytePrefixes = ['', 'kilo', 'mega', 'giga', 'tera', 'peta'];
+    const prefix = this.unit === 'bit' ? bitPrefixes : bytePrefixes;
+    const index = Math.max(0, Math.min(Math.floor(Math.log10(this.value) / 3), prefix.length - 1));
+    const unit = prefix[index] + this.unit;
+    const valueToFormat = parseFloat((this.value / Math.pow(1000, index)).toPrecision(3));
+
+    return this.localize.number(valueToFormat, {
+      style: 'unit',
+      unit,
+      unitDisplay: this.display
     });
   }
 }